Added getting item id

+++ Configuration/include/Configuration/configuration_keys.hpp
@@ -12,7 +12,7 @@ typedef uint32_t configuration_key_t;
 
 uint8_t get_data_storage_size(const configuration_key_t& configuration_key);
 uint8_t get_group_id(const configuration_key_t& configuration_key);
-uint8_t get_item_id(const configuration_key_t& configuration_key);
+uint16_t get_item_id(const configuration_key_t& configuration_key);
 
 } // namespace configuration
 } // namespace ublox
